resgroup: retire proposed columns and adjust default settings
In resource group all the group settings can be set at runtime, but some
of them can not take effect immediately, internally we have to store the
proposed setting and the in-use setting.  We used to store these two
values in the catalog table and show them in the gp_toolkit views.
However they are not very useful and the users may find them confusing.

Now we retired the proposed columns, the `value` columns will store the
user-specified settings directly.

Also adjust the default memory settings for default groups and newly
created groups.

Co-authored-by: Hubert Zhang <hzhang@pivotal.io>
Co-authored-by: Ning Yu <nyu@pivotal.io>