Fix read pg_distributedlog bug (#7889)
In the function 'DistributedLog_AdvanceOldestXmin', it reads pg_distributedlog,
updates 'DistributedLogShared->oldestXmin' and truncates the logs whose xid is
smaller than 'DistributedLogShared->oldestXmin'. This looks good because it's
serialized under the lock DistributedLogControlLock. The problem is the read
function 'SimpleLruReadPage' may release and re-acquire the lock, and another
caller of DistributedLog_AdvanceOldestXmin may truncate the logs concurrently
and 'SimpleLruReadPage' will error out. DistributedLog_CommittedCheck may
read LRU page in a similar way, also need to be protected.

The number of pg_distributedlog files is about 32 times of pg_clog, truncate
pg_distributedlog based on the frozenxid might left too many files on the disk.
So we introduce a new LWLock to prevent truncating a file while someone else is
reading it.

Co-authored-by: Gang Xiong <gxiong@pivotal.io>
Co-authored-by: Hubert Zhang <hzhang@pivotal.io>