resgroup: only sync memory_spill_ratio when necessary
The memory_spill_ratio can be set both via a catalog property or a GUC,
at beginning of each transaction the value will be synced from the
catalog to the GUC.  Setting a GUC is not too expansive, however if we
have to set it on each transaction then it's a significant overhead.

Optimize it by only sync it when the catalog and GUC have different
values.

(cherry picked from commit 1be1b802df899e9f4658d12b1fcb888c69102aaf)