Make fts_recovery_in_progress more robust
The test is found to fail intermittently in CI when the gprecoverseg command is
fired and a mirror that was promoted by FTS is still undergoing promotion.  Fix
this by running a distributed query in a loop until the query runs
successfully.

Reviewed by Jimmy Yih and Adam Berlin

(cherry picked from commit 71754f8bb1f84e1aeebe95fba4d3eee22b675dca)