Isolation2 framework: use utility mode connect when possible
When '<content_id>U:' is encountered in isolation2 spec, the framework opens a
connection to master and obtains hostname and port of the primary segment with
specified content ID.  This connection to master should be in utility mode.  A
normal mode connection would wait for DTM recovery and if a test is in the
middle of testing DTM recovery, the connection may not even complete.

Reviewed by Pengzhou Tang

(cherry picked from commit bbbe5da7222bcf7294b48e182953703562277edc)