Fix the race in commit_blocking_on_standby test properly
The test injects a skip fault on standby and then starts a create
table command in background.  The create table command is expected to
block due to the fault.  The test used to run the create table command
before waiting for the fault to be triggered.  Sometimes, the command
was found to complete without blocking if it reached faster than the
previously started fault injection command.

Make the test deterministic by ensuring that the skip fault is
triggered on standby and only then start the create table command.
Introduce wait loop until the create table command is shown as waiting
in pg_stat_activity.

Commit 7be7e1b3e9cc tried to fix this hastily.  This patch should fix
it properly.

Reviewed by Jimmy Yih

(cherry picked from commit b2756a1707444c0f47961a04424aa2344793e9cc)