Fix issue that GPDB auxiliary backend cannot start in QE
background worker is not scheduled until distributed transactions
are recovered if it needs to start at BgWorkerStart_RecoveryFinished
or BgWorkerStart_ConsistentState because it's not safe to do a read
or write if DTX is not recovered. GPDB is designed to do this check
in master only, however Gp_role == GP_ROLE_DISPATCH is not a
sufficient check for master.

Spotted by Wang Hao <haowang@pivotal.io>