Backport 6X: Re-enable gppkg behave test for both centos and ubuntu (#8067)
* Revert "CI: skip CLI behave tests that currently fail on ubuntu18.04"

This reverts commit 6a0cb2c601c902c5d417a1165e61645e3b612ccd.
This commit is to re-enable behave test for centos/ubuntu

* Ouput the same error message format

gppkg for rpm and deb outputs error messages when consistency is broken, but
the message is not the same, which make a pipeline fail.

* Update gppkg for behave test and fix some minor error in gppkg

Now, the binary sample.gppkg is removed, instead we add sample-*.rpm and sample-*.deb.
Because sample.gppkg is platform specific and GP major version sensitive. The uploaded
rpm/deb files are available unless the rpm/deb file is incompatible on the specific
platform. GP_MAJORVERSION is dynamically retrieved from a makefile installed in the gpdb
folder, so the gp major version in the gppkg will be always correct.
sample.gppkg will only be generated when gppkg tag is provided