Fix confusion between CURSOR_OPT_GENERIC_PLAN and CURSOR_OPT_UPDATABLE.
The PostgreSQL 9.2 merge introduced a couple of new cursor options,
including CURSOR_OPT_GENERIC_PLAN, which clashed with the 0x0040 value that
we were already using in GPDB for the GPDB-specific CURSOR_OPT_UPDATABLE
option. Fix the conflict.

The CURSOR_OPT_UPDATABLE flag was used in cursors opened in PL/pgSQL, so
the mixup caused generic plans to be used for all PL/pgSQL cursors. Add
a test case for that.

The PostgreSQL 9.2 merge happened in GPDB 6, so backport to GPDB 6.

Reviewed-by: Asim R P <apraveen@pivotal.io>
Reviewed-by: Pengzhou Tang <ptang@pivotal.io>