Allow concourse_unit_tests to run without .bats files present
When there are no bats testcases present, the current coding returned an
error due to the glob not being able to expand. This was made clear by
the recent removal of the last test, and the discussion in #8270. Rather
than using a glob, use ls and if statement. This allows us to keep the
unittest framework in place even if there are currently no tests.

Co-authored-by: Amil Khanzada <akhanzada@pivotal.io>
Co-authored-by: Daniel Gustafsson <dgustafsson@pivotal.io>
(cherry picked from commit fabc6ca5f6f747819fe850c071778ff80b4dcd75)