Print more details when a behave test fails
If a test decides to fail because not all segments are running, it will
be nice to get the details of segments that are found to be not running.

The details printed by this patch will only be shown by behave upon a test
failure.  They will not be shown when a test passes.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>
Reviewed-by: Jacob Champion <pchampion@pivotal.io>
Reviewed-by: David Krieger <dkrieger@pivotal.io>
(cherry picked from commit 2237ef21bb0f3a2ca8c30bd62c87dcf9c68b411e)