Error out when locallock and proclock are not consistent.
When acquire a lock, there is a case where locallock doesn't hold
the lock but global lock treats we already held the lock. This
inconsistent state may be caused by broken of locallock or
global lock.

Upstream elog error for this case from the begining. If we already
know locallock or global lock is broken, there is no reason to
continue the queries.