Use versioned_file for gp-internal-artifacts
The build artifacts in gp-internal-artifacts are not immutable. If
`regexp` is used when fetching the resource, then new builds (with the
same version) are not detected. This means that fixes in the packaging
may not be correctly picked up by a pipeline.

Authored-by: Bradford D. Boyle <bboyle@pivotal.io>
(cherry picked from commit 563e9f60069819bfe7b8231cff47f29a406a43d0)