set_config_by_name(): dispatch all values correctly
The dispatch code for set_config() incorrectly quoted the GUC value by
surrounding it in single quotes. Not only does this not handle more
"interesting" values containing embedded single quotes, but GUCs with
the GUC_LIST_QUOTE flag, such as search_path, ended up with different
values on the segments compared to the master. For example,

    SELECT set_config('search_path', 'my_schema,public', false);

was dispatched as

    SET search_path TO 'my_schema,public';

instead of

    SET search_path TO my_schema,public;

The two are not equivalent; the latter correctly sets search_path to a
list of two elements, "my_schema" and "public", whereas the former sets
search_path to a list containing a single "my_schema,public" element.

Rather than trying to parse apart the list value, dispatch the
set_config() call to the QEs as well. This lets us pass through the
(quoted) arguments directly and guarantees that we run the same code in
both places.

Co-authored-by: Shoaib Lari <slari@pivotal.io>
(cherry picked from commit c891c8ac8669c510e036d3987fe7e509f3f21f60)