Correctly use atomic variable in ResGroupControl.freeChunks. (#8434)
This variable was used mixing with atomic api functions and direct access.
This is not wrong usually in real scenario but is not a good implementation
since 1) that depends on compiler and H/W to ensure the correctness of direct
access. 2) code is not graceful.

Changing to all use atomic api functions.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>