6X: Avoid duplicated repeat node. (#8531)
Greenplum uses repeat node in plan to handle the case
that duplicated groups in grouping sets. Previously,
we add this node at two places:
  1. make_list_aggs_for_rollup, when is not twostage agg
  2. grouping_planner, when group-by keys are not collocated with dist-keys

But under some contitions, there are code paths that will add duplicated
repeat nodes. For example, when the GUC gp_eager_one_phase_agg is set to on,
, the grouping sets contains duplicated groups and group-by keys are not
collcated with dist-keys. In such case, repeate node is added both in 1 and
2.

This commit fixes the bug by:
  1. removing the code adding repeat node in grouping_planner
  2. let make_one_stage_agg_plan and make_two_stage_agg_plan handle adding
     repeat node

Co-authored-by: Zhenghua Lyu <zlv@pivotal.io>