Silece stringop-overflow warning and protect against invalid read (#8620)
strncat unfortunately takes a misleading size argument which means
at most size from src. It is a bit of an antipattern in the string
family of functions and for that compilers will emit a warning if
it happens that the size argument matches the size of src, since that
is not what usually users of strncat want to do.

The usage of the function in the code was correct. However instead
of silencing the compiler, strncat was replaced with a front end
dynamic buffer and family of functions that postgres is providing.

Also protect against an invalid read in case that the size of the
result is zero.