Fix build system to work with non-empty DESTDIR
When building gpdb with a non-empty DESTDIR, the build would fail
because certain Makefiles were not correctly account for it.
Additionally, when we make a symlink for `gpcheckcat` we *should not*
include DESTDIR in the target. A common usage for DESTDIR is to allow
moving the build artifacts after the build is completed. If the target
includes DESTDIR, than the link could point to a non-existing path.

Authored-by: Bradford D. Boyle <bboyle@pivotal.io>
(cherry picked from commit 7de35b7ee80da1c688378bbc53bd4eda72d786b8)