Upgrade tablespaces
The upgrade workflow for Greenplum works like this:

init-new-cluster -> upgrade-dispatcher -> copy-from-master ->
upgrade-segments

pg_upgrade is only responsible for the upgrade-dispatcher and
upgrade-segments steps, but it depends on the ordering of these steps.

This commit adds support for tablespaces during the copy-from-master step,
as well as adding logic inside of pg_upgrade necessary to map files
from GPDB5's filespaces to GPDB6's tablespaces. Described here:

During the upgrade-dispatcher step, gpupgrade can run pg_upgrade
--mode=dispatcher which will generate a file (old_tablespaces.txt)
into the current directory containing the tablespace information
inside of the GPDB6 cluster. This file can then be used to perform the
copy-from-master step and the upgrade-segments step.