Change fix for Motion node's tdhasoids confusion, and add tests.
In master, we already got this fix as part of the 9.5 merge. Change it
slightly, so that we don't use the subplan's tuple descriptor as is, but
only copy the "tdhasoid" flag from it. It's in principle possible that
some unimportant information, like attribute names or typmods, are not
set up in the subplan's result type, and should be computed from the
Motion's target list. We're not seeing problems like that on master, but
on the 9.6 merge branch we are.

More importantly, this adds bespoken tests for this scenario. It arose
in the 'rowsecurity' test, but this doesn't have anything to do with
row-level security, so it was accidental.

We didn't backport the fix to 6X_STABLE before, so do that now.

Fixes https://github.com/greenplum-db/gpdb/issues/8765.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>