Handle dispatching of rule statements
The common practice is to dispatch already-analysed statements. However for
rules raw statements where dispatched. As a consequence all the possible nodes
in the tree would have to be serialised. This was (and still) is not the case.

Follow the common practice and dispatch the already analysed query tree for
rules.

The issue rose during the 9.6 merge cycle in 'rowsecurity' tests by coincidence.
Rule tests are dissabled for quite some time now and even if this case was
included it was never tested. A specific test is added now.

Addresses issue #8787

Reported, reviewed and heavily modified by Heikki Linnakangas.