gpinitsystem: --ignore-warnings flag will return 0 on warnings
Previously gpinitsystem returned 1 in case of warnings (2 in case of
errors). We have changed gpinitsystem to return 0 for warnings so that
other tools using gpinitsystem (such as gpupgrade) will not fail when
gpinitsystem generates warnings. This behavior is only available when
specifying the --ignore-warnings flag. Existing exit code behavior
remains the same when the flag is not specified.

Also, gpinitsystem was exiting with non-zero return code if certain
errors or warnings had happened on any previous run of gpinitsystem that
day. This has also been fixed in this commit, in the case where
--ignore-warnings is specified.

Co-authored-by: Mark Sliva <msliva@pivotal.io>