Move server-build publishing to a separate CI job
The `without_asserts` pipeline includes a time-based trigger which can
cause the pipeline to re-run the compilation jobs for the same commit.
When the compilation jobs attempt to re-`put` the server-build artifact,
the job fails because the build artifacts bucket is immutable.

This commit move the server-build publishing to a separate, hidden CI
job that does not block the release.

Authored-by: Bradford D. Boyle <bboyle@pivotal.io>
(cherry picked from commit aac2eeb919aca8e118993ca712387fc879451599)