Don't create symbolic links to absolute paths
The DESTDIR would be determined while installation if with a tarball or
shell binary package, which is probably not the same as the DESTDIR we
used to make the tarball, in this case, the symbolic would link to a
non-existing or wrong file.

Use relative paths instead.

(cherry picked from commit ccb1e8ea37150b9c4f402776ce97bbce257368c9)