The PXF JDBC user identity in the documentation is incorrect for one
For one of the Hive authentication methods, the PXF documentation states
the incorrect identity used when Kerberos is enabled for Hive and the
hive.server2.enable.doAs property is false. The correct identity is the
Kerberos principal defined in the jdbc.url to connect to the Hive
database. Cherry-picked from 99784205b45216580e08d5beb5c28c92a8cf4020